Fix kodak camlib compilation warnings.
* dc120/dc120.c: Fix signed-unsigned comparison issue.
* dc120/library.c: Fix signed-unsigned comparison issue.
* dc210/dc210.c: Fix signed-unsigned comparison issues.
* dc240/library.c: Resolves statement fall through warning in switch block and couple signed-unsigned compare issues.
* dc3200/dc3200.c: Fix signed-unsigned comparison issues.
* ez200/ez200.c: Adds missing .idProduct initialization value.

+++ camlibs/kodak/dc120/dc120.c
@@ -143,7 +143,7 @@ static int find_folder( Camera *camera, const char *folder,
     for( i = 0; i<gp_list_count( albums ); i++ )
     {
 	gp_list_get_name( albums, i, &album_name );
-	if( strlen( album_name ) == folder_len &&
+	if( (int)strlen( album_name ) == folder_len &&
 	    strncmp( album_name, folder, folder_len ) == 0 )
 	{
 	    *folder_nr = i+1;
