+++ camlibs/kodak/dc210/dc210.c
@@ -173,7 +173,7 @@ camera_get_config (Camera *camera, CameraWidget **window, GPContext *context)
 	CameraWidget *section, *widget;
 	CameraAbilities abilities;
 	GPPortSettings settings;
-	int i;
+	unsigned int i;
 	char * wvalue;
 	char stringbuffer[12];
 
@@ -262,7 +262,7 @@ camera_get_config (Camera *camera, CameraWidget **window, GPContext *context)
         gp_widget_append (section, widget);
 	for (i = 0; i < sizeof(exp_comp)/sizeof(*exp_comp); i++){
 		gp_widget_add_choice (widget, exp_comp[i]);
-		if (status.exp_compensation + 4 == i)
+		if (status.exp_compensation + 4 == (int)i)
 			gp_widget_set_value (widget, exp_comp[i]);
 	};
 
@@ -353,7 +353,7 @@ camera_set_config (Camera *camera, CameraWidget *window, GPContext *context)
 {
 	CameraWidget *w, *w2;
 	char *wvalue, *w2value;
-	int i;
+	unsigned int i;
 
 	gp_widget_get_child_by_label (window, _("File type"), &w);
 	if (gp_widget_changed (w)) {
@@ -427,7 +427,7 @@ camera_set_config (Camera *camera, CameraWidget *window, GPContext *context)
 		gp_widget_get_value (w, &wvalue);
 		for (i = 0; i < sizeof(exp_comp)/sizeof(*exp_comp); i++){
 			if (strncmp(wvalue, exp_comp[i], 4) == 0){
-				dc210_set_exp_compensation(camera, i - 4);
+				dc210_set_exp_compensation(camera, (int)i - 4);
 				break;
 			};
 		};
